+++ smac/runner/__init__.py
@@ -1,9 +1,5 @@
 from smac.runner.abstract_runner import AbstractRunner
 from smac.runner.dask_runner import DaskParallelRunner
-from smac.runner.exceptions import (
-    FirstRunCrashedException,
-    TargetAlgorithmAbortException,
-)
 from smac.runner.target_function_runner import TargetFunctionRunner
 
 __all__ = [
@@ -11,7 +7,4 @@ __all__ = [
     "AbstractRunner",
     "TargetFunctionRunner",
     "DaskParallelRunner",
-    # Exceptions
-    "TargetAlgorithmAbortException",
-    "FirstRunCrashedException",
 ]